Update auth flow and update error handling.

+++ index.php
@@ -6,6 +6,13 @@ $app = new Silex\Application();
 
 // Debug helper provided by Silex
 $app['debug'] = TRUE;
+// TODO:
+//  Search messages
+//  If a notes folder doesn't exist create it to continue...
+//  OAuth with gmail (limit surface area)
+//  Simple session manager?
+//   - redirects on session
+//  Icon (white) on black navbar top
 
 // register the session extension
 $app->register(new Silex\Provider\SessionServiceProvider());
@@ -14,11 +21,65 @@ $app->register(new Silex\Provider\TwigServiceProvider(), array(
 ));
 
 $app->get('/', function() use($app) {
+    $app['session']->start();
+    $username = $app['session']->get('user');
+    if (!empty($username)) {
+        return $app->redirect('/note');
+    }
+    // if session just redirect to /note
     return $app['twig']->render('index.twig', array());
 });
 
+$app->match('/advauth', function(Symfony\Component\HttpFoundation\Request $request) use($app) {
+    $host = $request->get('host');
+    $user = $request->get('user');
+    $pass = $request->get('pass');
+    $port = $request->get('port');
+    $ssl = (bool) $request->get('ssl') == 'on';
+    $basefolder = $request->get('basefolder');
+
+    if (!empty($host) && !empty($user)) {
+        $app['session']->start();
+        $app['session']->set('host', $host);
+        $app['session']->set('user', $user);
+        $app['session']->set('pass', $pass);
+        $app['session']->set('port', (int) $port);
+        $app['session']->set('ssl', (bool) $ssl);
+        $app['session']->set('basefolder', empty($basefolder) ? NULL : $basefolder);
+        $notes = notesSessionMgr($app);
+        return $app->redirect('/note');
+    }
+
+    return $app['twig']->render('advauth.twig', array());
+})->method('GET|POST');
+
+function notesSessionMgr($app) {
+    $app['session']->start();
+    $host = $app['session']->get('host');
+    $user = $app['session']->get('user');
+    $pass = $app['session']->get('pass');
+    $port = (int) $app['session']->get('port');
+    $ssl = (bool) $app['session']->get('ssl');
+    $basefolder = $app['session']->get('basefolder');
+    $basefolder = empty($basefolder) ? NULL : $basefolder;
+    try {
+        $notes = new Notes($host, $user, $pass, $basefolder, $port, $ssl);
+        $notes->count();
+        return $notes;
+    }
+    catch (Exception $e) {
+        $app['session']->clear();
+        //return $app->abort(300, 'Uh Oh - going back home', array("Location" => "http://hackmw.dev/"));
+        return $app->redirect('/');
+    }
+}
+
+// ////////// Note app
 $app->get('/note', function() use($app) {
-    $notes = new Notes('mail.messagingengine.com', 'minenet@airpost.net', 'P@ssw0rd', 'INBOX');
+    // $notes = new Notes('mail.messagingengine.com', 'minenet@airpost.net', 'P@ssw0rd', 'INBOX');
+    $notes = notesSessionMgr($app);
+    if (strpos(get_class($notes), 'Notes') === FALSE) return $app->redirect('/');
+
     // TODO: Move these to tests
     $index = $notes->index();
 
@@ -28,7 +89,8 @@ $app->get('/note', function() use($app) {
 });
 
 $app->get('/note/{id}', function($id) use($app) {
-    $notes = new Notes('mail.messagingengine.com', 'minenet@airpost.net', 'P@ssw0rd', 'INBOX');
+    $notes = notesSessionMgr($app);
+    if (strpos(get_class($notes), 'Notes') === FALSE) return $app->redirect('/');
     $add = FALSE;
     // TODO: Move these to tests
     if (strpos($id, 'add') !== FALSE) {
@@ -59,7 +121,8 @@ $app->get('/note/{id}', function($id) use($app) {
 });
 
 $app->post('/note/{id}', function(Symfony\Component\HttpFoundation\Request $request, $id) use($app) {
-    $notes = new Notes('mail.messagingengine.com', 'minenet@airpost.net', 'P@ssw0rd', 'INBOX');
+    $notes = notesSessionMgr($app);
+    if (strpos(get_class($notes), 'Notes') === FALSE) return $app->redirect('/');
     $save = $request->get('save');
     $delete = $request->get('delete');
     if (!empty($save)) {
@@ -80,22 +143,10 @@ $app->post('/note/{id}', function(Symfony\Component\HttpFoundation\Request $requ
 });
 
 $app->delete('/note/{id}', function($id) use($app) {
-    $notes = new Notes('mail.messagingengine.com', 'minenet@airpost.net', 'P@ssw0rd', 'INBOX');
+    $notes = notesSessionMgr($app);
+    if (strpos(get_class($notes), 'Notes') === FALSE) return $app->redirect('/');
     $notes->delete($id);
     return new Symfony\Component\HttpFoundation\Response('Deleted', 201);
-});
-
-$app->get('/email', function() use($app) {
-    $app['session']->start();
-    $notes = new Notes('mail.messagingengine.com', 'minenet@airpost.net', 'P@ssw0rd', 'INBOX');
-    // TODO: Move these to tests
-    $debug = $notes->index();
-    //$debug = $notes->create(rand(0, 1000) . "Hello There\n\n\n World One \n Two Three");
-    //$debug = $notes->create("uh oh");
-    //$debug = $notes->retrieve('4adb0460-acf5-11e1-81f0-7f6c2cc069f2');
-    //$debug = $notes->retrieve(2);
-    //$debug = $debug = $notes->delete('410e98c0-acf5-11e1-a4c8-17f957629924');
-    print_r($debug);
 });
 
 $app->run();
